Fix sign

+++ src/decio.cxx
@@ -292,8 +292,8 @@ CDecor::Read (
   }
 
   /* Restore the flagged state where flagged ground */
-  for (int i = 0; i < countof (m_decor); ++i)
-    for (int j = 0; j < countof (m_decor[i]); ++j)
+  for (size_t i = 0; i < countof (m_decor); ++i)
+    for (size_t j = 0; j < countof (m_decor[i]); ++j)
     {
       if (m_decor[i][j].objectIcon == 124)
         m_decorMem[i][j].flagged = true;
